// old-pixi/src/types/grid.ts
export type Point = [number, number];

/**
 * Describes the type of cell, either a land cell for a landmass or a water cells from the water next to a landmass.
 */
export enum CellType {
  Highland = 2,
  Land = 1,
  Water = -1,
}
// ...
export interface Cells {
  /**
   * Array of indexes for each cell in the grid.
   */
  indexes: Uint32Array | Uint8Array | Uint16Array;

  /**
   * Cell vertices for the cell at the given index.
   */
  vertices: number[][];

  /**
   * Cells adjacent to the cell at the given index.
   */
  adjacentCells: number[][];

  /**
   * List of all the cells, identifying which one is a near border cell or not.
   */
  nearBorderCells: boolean[];

  /**
   * Array of heights from the heightmap for each cell.
   */
  heights: Uint32Array | Uint8Array | Uint16Array;

  /**
   * Array of features indexes for individual cells, can be used to query the actual feature from the features array of
   * the grid.
   */
  features: Uint16Array;

  /**
   * Array of types for cells to identify which cells are part of the land/coast of a landmass and which cells are the
   * water near the landmass.
   */
  types: CellType[];

  /**
   * Array of temperatures for the given cell IDs, generated by creating a random position for the map on a globe.
   */
  temperatures: Int8Array;

  /**
   * Array of precipitation for the given cell IDs, used to generate biomes for individual cells as it tells how dry or
   * wet an area of the map is.
   */
  precipitation: Uint8Array;

  /**
   * Array of water flux values for each cell, useful to generate rivers and biomes, in m³ per second.
   */
  waterFlux: Uint16Array;

  /**
   * Array of river indexes for a cell. 0 if there are no rivers on that cell, otherwise it defines the ID for the river
   * from the grid object.
   */
  rivers: Uint16Array;

  /**
   * Array of confluence for rivers, this describes when two or more rivers merge into one another.
   */
  confluences: Uint8Array | Uint16Array;
}
// ...
export interface Vertices {
  /**
   * Coordinates for the vertices of the cells at the given index, will be set to the center of the triangle generated
   * by delaunator.
   */
  coordinates: Point[];

  /**
   * Neighbouring vertices for the vertex at the given index.
   */
  neighbours: number[][];

  /**
   * Adjacent cells of the vertex at the given index.
   */
  adjacent: [number, number, number][];
}